requirements.txt
aiohttp
pandas
requests
ta
python-dotenv
pytz
websockets